Externalize CLI formatting to `explcheck-format.lua`
